完善核心报文，添加计数输入流

+++ livemq-core/src/main/java/org/livemq/core/wire/MqttAck.java
@@ -1,10 +1,13 @@
 package org.livemq.core.wire;
 
 /**
- * <h1>控制报文确认报文顶级父类</h1>
  * 
+ * @Title MqttAck
+ * @Package org.livemq.core.wire
+ * @Description 控制报文确认报文顶级父类
  * @author xinxisimple@163.com
- * @date 2018-07-04 13:40
+ * @date 2018-07-18 16:57
+ * @version 1.0.0
  */
 public abstract class MqttAck extends MqttWireMessage {
 
@@ -14,7 +17,7 @@ public abstract class MqttAck extends MqttWireMessage {
 
 	@Override
 	public byte getMessageInfo() {
-		return 0;
+		return (byte) 0;
 	}
 
 }
