+++ livemq-core/src/main/java/org/livemq/core/wire/MqttPingReq.java
@@ -1,32 +1,20 @@
 package org.livemq.core.wire;
 
 /**
- * <h1>控制报文 - PINGREQ</h1>
  * 
- * <p>
- * 心跳检测报文 <br>
- * 
- * <h2>客户端发送 PINGREQ 报文给服务端。用于：</h2>
- * <ol>
- * 	<li>在没有任何其它控制报文从客户端发送给服务端时，告知服务端客户端还活着。</li>
- * 	<li>请求服务端发送 响应确认它还活着。</li>
- * 	<li>使用网络以确认网络连接没有断开。</li>
- * </ol>
- * 
- * 保持连接（Keep Alive）处理中用到这个报文，详情信息请查看 3.1.2.10 节。<br>
- * 
- * <h2>可变报头</h2>
- * PINGREQ 报文没有可变报头。<br>
- * 
- * <h2>有效荷载</h2>
- * PINGREQ 报文没有有效荷载。<br>
- * 
- * <h2>响应</h2>
- * 服务端必须发送 PINGRESP 报文响应客户端的 PINGREQ 报文。<br>
- * </p>
- * 
+ * @Title PINGREQ - 心跳请求
+ * @Package org.livemq.core.wire
+ * @Description 客户端发送 PINGREQ 报文给服务端。用于：
+ * 		<ol>
+ * 			<li>在没有任何其它控制报文从客户端发给服务端时，告知服务端客户端还活着。</li>
+ * 			<li>请求服务端发送，响应确认它还活着。</li>
+ * 			<li>使用网络以确认网络连接没有断开。</li>
+ * 		</ol>
+ * 		保持连接（Kepp Alive）处理中用到这个报文。
  * @author xinxisimple@163.com
- * @date 2018-07-04 14:10
+ * @date 2018-07-18 16:50
+ * @version 1.0.0
+ * @see https://github.com/mcxiaoke/mqtt/blob/master/mqtt/0312-PINGREQ.md
  */
 public class MqttPingReq extends MqttWireMessage {
 
@@ -45,7 +33,7 @@ public class MqttPingReq extends MqttWireMessage {
 
 	@Override
 	public byte getMessageInfo() {
-		return 0;
+		return (byte) 0;
 	}
 	
 	@Override
